adding functionality to get true/false for checking if a host is a config server
